Merging feature/auction into develop

+++ src/controllers/helpers/DatabaseHelper.ts
@@ -480,7 +480,7 @@ const DatabaseHelper = {
 	  
 	  return RelationalDatabaseORMClient.models[schema.group];
 	},
-	insert: async (data: Input[], baseSchema: DataTableSchema, crossRelationUpsert: boolean=false): Promise<HierarchicalDataRow[]> => {
+	insert: async (data: Input[], baseSchema: DataTableSchema, crossRelationUpsert: boolean=false, session: any=null): Promise<HierarchicalDataRow[]> => {
 		return new Promise(async (resolve, reject) => {
   		const transaction = await CreateTransaction({});
   		
@@ -492,22 +492,22 @@ const DatabaseHelper = {
   		  	const input = list[index][0];
   		  	const schema = list[index][1];
   		  	
-  		  	await DatabaseHelper.performRecursiveInsert(input, schema, results, transaction, crossRelationUpsert);
+  		  	await DatabaseHelper.performRecursiveInsert(input, schema, results, transaction, crossRelationUpsert, session);
   		  }
 	      
       	if (transaction) await transaction.commit();
 		  	
 	  		resolve(results);
       } catch(error) {
+      	console.log(error);
+      	
       	if (transaction) await transaction.rollback();
       	
         reject(error);
       }
     });
 	},
-	performRecursiveInsert: async (input: HierarchicalDataTable, schema: DataTableSchema, results: HierarchicalDataRow[], transaction: any, crossRelationUpsert: boolean=false) => {
-    if (!PermissionHelper.validate(ActionType.Insert, schema)) throw new Error(`You have no permission to insert any row in ${schema.group}.`);
-    
+	performRecursiveInsert: async (input: HierarchicalDataTable, schema: DataTableSchema, results: HierarchicalDataRow[], transaction: any, crossRelationUpsert: boolean=false, session: any=null) => {
     switch (input.source) {
     	case SourceType.Relational:
     		if (!RelationalDatabaseClient) throw new Error("There was an error trying to obtain a connection (not found).");
@@ -531,6 +531,8 @@ const DatabaseHelper = {
 					    }
 					  }
 					}
+					
+					if (await !PermissionHelper.allowActionOnTable(ActionType.Insert, schema, hash, session)) throw new Error(`You have no permission to insert any row in ${schema.group}.`);
 					
 					const record = await map.create(hash, {transaction: transaction});
 					
@@ -580,19 +582,19 @@ const DatabaseHelper = {
 							  rows: []
 						  };
 						
-							if (!crossRelationUpsert) await DatabaseHelper.performRecursiveInsert(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction);
-							else await DatabaseHelper.performRecursiveUpsert(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction);
+							if (!crossRelationUpsert) await DatabaseHelper.performRecursiveInsert(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction, session);
+							else await DatabaseHelper.performRecursiveUpsert(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction, session);
 						}
 					}
 				  
 				  for (const key in schema.columns) {
 					  if (schema.columns.hasOwnProperty(key) && result.columns[key] !== undefined) {
-					    if (!PermissionHelper.allow(schema.columns[key])) delete result.columns[key];
+					    if (await !PermissionHelper.allowOutputOfColumn(schema.columns[key], schema, session)) delete result.columns[key];
 					  }
 					}
 					for (const key in schema.keys) {
 					  if (schema.keys.hasOwnProperty(key) && result.keys[key] !== undefined) {
-					    if (!PermissionHelper.allow(schema.keys[key])) delete result.keys[key];
+					    if (await !PermissionHelper.allowOutputOfColumn(schema.keys[key], schema, session)) delete result.keys[key];
 					  }
 					}
 				}
@@ -617,9 +619,7 @@ const DatabaseHelper = {
     		break;
     }
 	},
-	performRecursiveUpsert: async (input: HierarchicalDataTable, schema: DataTableSchema, results: HierarchicalDataRow[], transaction: any) => {
-		if (!PermissionHelper.validate(ActionType.Insert, schema)) throw new Error(`You have no permission to upsert any row in ${schema.group}.`);
-		
+	performRecursiveUpsert: async (input: HierarchicalDataTable, schema: DataTableSchema, results: HierarchicalDataRow[], transaction: any, session: any=null) => {
     switch (input.source) {
     	case SourceType.Relational:
     		if (!RelationalDatabaseClient) throw new Error("There was an error trying to obtain a connection (not found).");
@@ -643,6 +643,8 @@ const DatabaseHelper = {
 					    }
 					  }
 					}
+					
+					if (await !PermissionHelper.allowActionOnTable(ActionType.Insert, schema, hash, session)) throw new Error(`You have no permission to upsert any row in ${schema.group}.`);
 					
 					const record = await map.upsert(hash, {transaction: transaction});
 					
@@ -692,18 +694,18 @@ const DatabaseHelper = {
 							  rows: []
 						  };
 						
-							await DatabaseHelper.performRecursiveUpsert(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction);
+							await DatabaseHelper.performRecursiveUpsert(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction, session);
 						}
 					}
 				
 				  for (const key in schema.columns) {
 					  if (schema.columns.hasOwnProperty(key) && result.columns[key] !== undefined) {
-					    if (!PermissionHelper.allow(schema.columns[key])) delete result.columns[key];
+					    if (await !PermissionHelper.allowOutputOfColumn(schema.columns[key], schema, session)) delete result.columns[key];
 					  }
 					}
 					for (const key in schema.keys) {
 					  if (schema.keys.hasOwnProperty(key) && result.keys[key] !== undefined) {
-					    if (!PermissionHelper.allow(schema.keys[key])) delete result.keys[key];
+					    if (await !PermissionHelper.allowOutputOfColumn(schema.keys[key], schema, session)) delete result.keys[key];
 					  }
 					}
 				}
@@ -728,7 +730,7 @@ const DatabaseHelper = {
     		break;
     }
 	},
-	update: async (data: Input[], baseSchema: DataTableSchema, crossRelationUpsert: boolean=false): Promise<HierarchicalDataRow[]> => {
+	update: async (data: Input[], baseSchema: DataTableSchema, crossRelationUpsert: boolean=false, session: any=null): Promise<HierarchicalDataRow[]> => {
 		return new Promise(async (resolve, reject) => {
   		const transaction = await CreateTransaction({});
   		
@@ -740,22 +742,22 @@ const DatabaseHelper = {
   		  	const input = list[index][0];
   		  	const schema = list[index][1];
   		  	
-  		  	await DatabaseHelper.performRecursiveUpdate(input, schema, results, transaction, crossRelationUpsert);
+  		  	await DatabaseHelper.performRecursiveUpdate(input, schema, results, transaction, crossRelationUpsert, session);
   		  }
 	      
       	if (transaction) await transaction.commit();
 		  	
 	  		resolve(results);
       } catch(error) {
+      	console.log(error);
+      	
       	if (transaction) await transaction.rollback();
       	
         reject(error);
       }
     });
 	},
-	performRecursiveUpdate: async (input: HierarchicalDataTable, schema: DataTableSchema, results: HierarchicalDataRow[], transaction: any, crossRelationUpsert: boolean=false) => {
-		if (!PermissionHelper.validate(ActionType.Update, schema)) throw new Error(`You have no permission to update any row in ${schema.group}.`);
-		
+	performRecursiveUpdate: async (input: HierarchicalDataTable, schema: DataTableSchema, results: HierarchicalDataRow[], transaction: any, crossRelationUpsert: boolean=false, session: any=null) => {
     switch (input.source) {
     	case SourceType.Relational:
     		if (!RelationalDatabaseClient) throw new Error("There was an error trying to obtain a connection (not found).");
@@ -776,6 +778,8 @@ const DatabaseHelper = {
 					    data[key] = row.columns[key];
 					  }
 					}
+					
+					if (await !PermissionHelper.allowActionOnTable(ActionType.Update, schema, hash, session)) throw new Error(`You have no permission to update any row in ${schema.group}.`);
 					
 					await map.update(data, {where: hash, transaction: transaction});
 					
@@ -826,18 +830,18 @@ const DatabaseHelper = {
 							  rows: []
 						  };
 						
-							await DatabaseHelper.performRecursiveUpsert(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction);
+							await DatabaseHelper.performRecursiveUpsert(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction, session);
 						}
 					}
 				
 				  for (const key in schema.columns) {
 					  if (schema.columns.hasOwnProperty(key) && result.columns[key] !== undefined) {
-					    if (!PermissionHelper.allow(schema.columns[key])) delete result.columns[key];
+					    if (await !PermissionHelper.allowOutputOfColumn(schema.columns[key], schema, session)) delete result.columns[key];
 					  }
 					}
 					for (const key in schema.keys) {
 					  if (schema.keys.hasOwnProperty(key) && result.keys[key] !== undefined) {
-					    if (!PermissionHelper.allow(schema.keys[key])) delete result.keys[key];
+					    if (await !PermissionHelper.allowOutputOfColumn(schema.keys[key], schema, session)) delete result.keys[key];
 					  }
 					}
 				}
@@ -862,7 +866,7 @@ const DatabaseHelper = {
     		break;
     }
   },
-	retrieve: async (data: Input[], baseSchema: DataTableSchema): Promise<{[Identifier: string]: HierarchicalDataTable}> => {
+	retrieve: async (data: Input[], baseSchema: DataTableSchema, session: any=null): Promise<{[Identifier: string]: HierarchicalDataTable}> => {
 		return new Promise(async (resolve, reject) => {
 		  try {
 		  	if (data != null) {
@@ -873,12 +877,12 @@ const DatabaseHelper = {
 	  		  	const input = list[index][0];
 	  		  	const schema = list[index][1];
 	  		  	
-	  		  	await DatabaseHelper.performRecursiveRetrieve(input, schema, results);
+	  		  	await DatabaseHelper.performRecursiveRetrieve(input, schema, results, session);
 	  		  }
 			  	
 		  		resolve(results);
 		  	} else {
-		  		if (!PermissionHelper.validate(ActionType.Retrieve, baseSchema)) throw new Error(`You have no permission to retrieve any row in ${baseSchema.group}.`);
+		  		if (await !PermissionHelper.allowActionOnTable(ActionType.Retrieve, baseSchema, {}, session)) throw new Error(`You have no permission to retrieve any row in ${baseSchema.group}.`);
 		  		
 		  		const results = {};
 		  		
@@ -901,12 +905,12 @@ const DatabaseHelper = {
 	  				  
 	  					  for (const key in baseSchema.columns) {
 	    					  if (baseSchema.columns.hasOwnProperty(key) && row.columns[key] !== undefined) {
-	    					    if (!PermissionHelper.allow(baseSchema.columns[key])) delete row.columns[key];
+	    					    if (await !PermissionHelper.allowOutputOfColumn(baseSchema.columns[key], baseSchema, session)) delete row.columns[key];
 	    					  }
 	    					}
 	    					for (const key in baseSchema.keys) {
 	    					  if (baseSchema.keys.hasOwnProperty(key) && row.keys[key] !== undefined) {
-	    					    if (!PermissionHelper.allow(baseSchema.keys[key])) delete row.keys[key];
+	    					    if (await !PermissionHelper.allowOutputOfColumn(baseSchema.keys[key], baseSchema, session)) delete row.keys[key];
 	    					  }
 	    					}
 	    					
@@ -943,13 +947,13 @@ const DatabaseHelper = {
 	        resolve(results);
 	      }
       } catch(error) {
+      	console.log(error);
+      	
         reject(error);
       }
 		});
 	},
-	performRecursiveRetrieve: async (input: HierarchicalDataTable, baseSchema: DataTableSchema, results: {[Identifier: string]: HierarchicalDataTable}) => {
-		if (!PermissionHelper.validate(ActionType.Retrieve, baseSchema)) throw new Error(`You have no permission to retrieve any row in ${baseSchema.group}.`);
-		
+	performRecursiveRetrieve: async (input: HierarchicalDataTable, baseSchema: DataTableSchema, results: {[Identifier: string]: HierarchicalDataTable}, session: any=null) => {
     switch (input.source) {
     	case SourceType.Relational:
     		if (!RelationalDatabaseClient) throw new Error("There was an error trying to obtain a connection (not found).");
@@ -969,6 +973,8 @@ const DatabaseHelper = {
 					    hash[key] = row.keys[key];
 					  }
 					}
+					
+					if (await !PermissionHelper.allowActionOnTable(ActionType.Retrieve, baseSchema, hash, session)) throw new Error(`You have no permission to retrieve any row in ${baseSchema.group}.`);
 					
 					const rows = [];
 					const records = await map.findAll({where: hash}) || [];
@@ -1022,7 +1028,7 @@ const DatabaseHelper = {
 					  			}
 					  		}
 							  
-							  await DatabaseHelper.performRecursiveRetrieve(row.relations[key], nextSchema, _row.relations);
+							  await DatabaseHelper.performRecursiveRetrieve(row.relations[key], nextSchema, _row.relations, session);
 					  	}
 					  }
 					}
@@ -1030,12 +1036,12 @@ const DatabaseHelper = {
 					for (const row of rows) {
 					  for (const key in baseSchema.columns) {
 						  if (baseSchema.columns.hasOwnProperty(key) && row.columns[key] !== undefined) {
-						    if (!PermissionHelper.allow(baseSchema.columns[key])) delete row.columns[key];
+						    if (await !PermissionHelper.allowOutputOfColumn(baseSchema.columns[key], baseSchema, session)) delete row.columns[key];
 						  }
 						}
 						for (const key in baseSchema.keys) {
 						  if (baseSchema.keys.hasOwnProperty(key) && row.keys[key] !== undefined) {
-						    if (!PermissionHelper.allow(baseSchema.keys[key])) delete row.keys[key];
+						    if (await !PermissionHelper.allowOutputOfColumn(baseSchema.keys[key], baseSchema, session)) delete row.keys[key];
 						  }
 						}
 					}
@@ -1061,7 +1067,7 @@ const DatabaseHelper = {
     		break;
     }
   },
-	delete: async (data: Input[], baseSchema: DataTableSchema): Promise<HierarchicalDataRow[]> => {
+	delete: async (data: Input[], baseSchema: DataTableSchema, session: any=null): Promise<HierarchicalDataRow[]> => {
 		return new Promise(async (resolve, reject) => {
   		const transaction = await CreateTransaction({});
   		
@@ -1073,22 +1079,22 @@ const DatabaseHelper = {
   		  	const input = list[index][0];
   		  	const schema = list[index][1];
   		  	
-  		  	await DatabaseHelper.performRecursiveDelete(input, schema, results, transaction);
+  		  	await DatabaseHelper.performRecursiveDelete(input, schema, results, transaction, session);
   		  }
 	      
       	if (transaction) await transaction.commit();
 	      
 	      resolve(results);
       } catch(error) {
+      	console.log(error);
+      	
       	if (transaction) await transaction.rollback();
       	
         reject(error);
       }
     });
 	},
-	performRecursiveDelete: async (input: HierarchicalDataTable, schema: DataTableSchema, results: HierarchicalDataRow[], transaction: any) => {
-		if (!PermissionHelper.validate(ActionType.Delete, schema)) throw new Error(`You have no permission to delete any row in ${schema.group}.`);
-		
+	performRecursiveDelete: async (input: HierarchicalDataTable, schema: DataTableSchema, results: HierarchicalDataRow[], transaction: any, session: any=null) => {
     switch (input.source) {
     	case SourceType.Relational:
     		if (!RelationalDatabaseClient) throw new Error("There was an error trying to obtain a connection (not found).");
@@ -1099,10 +1105,12 @@ const DatabaseHelper = {
 					const hash = {};
 				
 					for (const key in schema.keys) {
-					  if (schema.keys.hasOwnProperty(key) && input.rows[0].keys[key] != undefined) {
-					    hash[key] = input.rows[0].keys[key];
+					  if (schema.keys.hasOwnProperty(key) && row.keys[key] != undefined) {
+					    hash[key] = row.keys[key];
 					  }
 					}
+					
+					if (await !PermissionHelper.allowActionOnTable(ActionType.Delete, schema, hash, session)) throw new Error(`You have no permission to delete any row in ${schema.group}.`);
 					
 					const record = await map.findOne({where: hash});
 					await record.destroy({force: true, transaction: transaction});
@@ -1154,18 +1162,18 @@ const DatabaseHelper = {
 							  rows: []
 						  };
 						
-							await DatabaseHelper.performRecursiveDelete(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction);
+							await DatabaseHelper.performRecursiveDelete(row.relations[key], nextSchema, result.relations[nextSchema.group].rows, transaction, session);
 						}
 					}
 				
 				  for (const key in schema.columns) {
 					  if (schema.columns.hasOwnProperty(key) && result.columns[key] !== undefined) {
-					    if (!PermissionHelper.allow(schema.columns[key])) delete result.columns[key];
+					    if (await !PermissionHelper.allowOutputOfColumn(schema.columns[key], schema, session)) delete result.columns[key];
 					  }
 					}
 					for (const key in schema.keys) {
 					  if (schema.keys.hasOwnProperty(key) && result.keys[key] !== undefined) {
-					    if (!PermissionHelper.allow(schema.keys[key])) delete result.keys[key];
+					    if (await !PermissionHelper.allowOutputOfColumn(schema.keys[key], schema, session)) delete result.keys[key];
 					  }
 					}
 				}
